+++ src/main/java/com/solutec/desarrollo_web_server/entity/RouteAssignment.java
@@ -13,7 +13,7 @@ public class RouteAssignment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "assignmentID")
-    private Integer assignmentID;
+    private Long assignmentID;
 
     @ManyToOne
     @JoinColumn(name = "routeId")
